+++ extras/display_code/doubles.py
@@ -1,0 +1,18 @@
+# Doubles, next two lines should be one
+
+
+class Thing:
+    bar = 1
+
+def foo():
+    print()
+    # next two lines should be one if trim_whitespace 
+
+    
+def do_nothing():
+    pass
+    # next two lines should be one if trim_whitespace 
+    
+    
+def nothing_else():
+    pass